dpp_dps: Add tests. Fix error in dpp_to_dps"
